rsk/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-ckeditor==6.1.0
django-environ==0.6.0
django-js-asset==1.2.2
django-request==1.5.6
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
